requirements.txt
clingo
gurobipy
tensorflow
keras
matplotlib
numpy
pandas
scikit_learn
scipy
seaborn
tabulate
torch
tqdm
pyyaml
wandb
scipy
lime
ucimlrepo
PyYAML
git+https://github.com/BogdanTurbal/nam.git